programm/model.py
"""
Модель для алгоритмов анализа спутниковых снимков
"""

import os
import cv2
import numpy as np
from PIL import Image

class Model:
    """Базовая модель для обработки изображений"""
    
    def __init__(self):
        self.weights_loaded = False
        
    def load_weights(self):
        """Загрузка весов модели"""
        self.weights_loaded = True
        print("Веса модели загружены")
        
    def detect_objects(self, image_path, confidence=0.5):
        """Детекция объектов на изображении"""
        # Здесь должна быть реальная логика детекции
        # Пока что возвращаем демонстрационные результаты
        return {
            'objects': [
                {'class': 'Корабль', 'count': 47},
                {'class': 'Гавань', 'count': 3},
                {'class': 'Теннисный корт', 'count': 2}
            ],
            'image_with_boxes': image_path
        }
        
    def analyze_roi(self, image_path, roi_type="автоматическое", sensitivity=50):
        """Анализ зон интересов"""
        # Здесь должна быть реальная логика анализа ROI
        return {
            'roi_regions': [
                {'type': 'Портовая зона', 'confidence': 0.95},
                {'type': 'Жилая зона', 'confidence': 0.87},
                {'type': 'Промышленная зона', 'confidence': 0.72}
            ],
            'roi_image': image_path
        }
        
    def enhance_image(self, image_path, enhance_type="автоматическое", intensity=50):
        """Улучшение качества изображения"""
        # Здесь должна быть реальная логика улучшения
        return {
            'enhanced_image': image_path,
            'quality_metrics': {
                'contrast': 0.85,
                'brightness': 0.78,
                'sharpness': 0.92
            }
        }
