programm/requirements.txt
PyQt5>=5.15.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.3.0
scikit-learn>=1.0.0
torch>=1.9.0
torchvision>=0.10.0
tensorflow>=2.6.0
requests>=2.25.0
